refactor(db): re-export starter db bootstrap and share isUniqueConflict
Re-export the package's lazy `db` proxy and `createDb` from
@naeemba/next-starter/db instead of hand-rolling a second postgres()+drizzle()
bootstrap, so the app honours the package's env-driven pool tuning and the two
can't drift. Extract the duplicated isUniqueConflict helper into
lib/db/isUniqueConflict.ts and drop the dead SQLite branch.
